added code to return user profile with requests

+++ src/services/mentorshipRequests.ts
@@ -1,32 +1,30 @@
 import { RequestStatus } from "@prisma/client";
 import { prisma } from "../config/prisma";
-import { createMentorshipRequest, MentorshipRequest } from "../types/user.types";
+import { createMentorshipRequest, MentorshipRequest } from "../types/user.types.js";
+import { NotFoundError } from "../errors/customErrors";
 
 export const createRequest = async (data: createMentorshipRequest): Promise<MentorshipRequest> => {
-  const dateTime = new Date(`${data.date}T${data.time}Z`); // Combine date and time into a Date object
   const request = await prisma.mentorshipRequest.create({
     data: {
       ...data,
-      date: dateTime,
     },
+    include: { mentor: true }
   });
   return request;
 };
 
 export const getSentRequests = async (menteeId: number): Promise<MentorshipRequest[]> => {
   const requests = await prisma.mentorshipRequest.findMany({
-    where: {
-      menteeId: menteeId,
-    },
+    where: { menteeId },
+    include: { mentor: true }
   });
   return requests;
 }
 
 export const getReceivedRequests = async (mentorId: number): Promise<MentorshipRequest[]> => {
   const requests = await prisma.mentorshipRequest.findMany({
-    where: {
-      mentorId: mentorId,
-    },
+    where: { mentorId },
+    include: { mentee: true}
   });
   return requests;
 };
@@ -35,11 +33,18 @@ export const updateRequestStatus = async (requestId: number, status: RequestStat
   const updatedRequest = await prisma.mentorshipRequest.update({
     where: { id: requestId },
     data: { status },
+    include: {mentee: true}
   });
   return updatedRequest;
 };
 
 export const deleteRequest = async (requestId: number): Promise<void> => {
+  const request = await prisma.mentorshipRequest.findUnique({
+    where: {id: requestId},
+  })
+  if (!request) {
+    throw new NotFoundError("Request not found")
+  }
   await prisma.mentorshipRequest.delete({
     where: { id: requestId },
   });
